+++ src/java/Breakout/src/main/java/breakout/app/GameObjects/SizeIncrease.java
@@ -1,11 +1,11 @@
 package breakout.app.GameObjects;
 
-
+ 
 import org.json.JSONObject;
 public class SizeIncrease extends PowerUp {
-    public float size_multiplier;
+    public int size_multiplier;
 
-    public SizeIncrease(float multiplier){
+    public SizeIncrease(int multiplier){
         this.size_multiplier = multiplier;
         this.description = new JSONObject();
         this.description.put("category", "racket");
