+++ src/java/Breakout/src/main/java/breakout/app/GameObjects/SpeedIncrease.java
@@ -1,11 +1,11 @@
 package breakout.app.GameObjects;
 
-import org.json.JSONObject;
+import org.json.JSONObject; 
 
 public class SpeedIncrease extends PowerUp {
-    public float speed_multipler;
+    public int speed_multipler;
 
-    public SpeedIncrease(float multiplier, String target){
+    public SpeedIncrease(int multiplier, String target){
         this.speed_multipler = multiplier;
         this.description = new JSONObject();
         this.description.put("category", target); // El target debe ser 'ball' o 'racket'
